moved more otel code to bsb-otel
